requirements.txt
Django>=5.2.4
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
ruff>=0.8.0
Pillow>=10.0.0

# AI Features - RAG & Vector Search
sentence-transformers>=2.2.0
torch>=2.0.0
pgvector>=0.2.0
google-generativeai>=0.3.0

# Document Processing (Phase 1)
pypdf2>=3.0.0
python-docx>=1.0.0
markdown>=3.5.0

# LangChain (Phase 2/3)
langchain==1.1.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.3
langchain-openai==1.1.2
langchain-text-splitters==1.0.0
openai==2.11.0
